+++ doc/sphinx/changelog.rst
@@ -18,6 +18,8 @@ performed stress tests and code audit DAS servers.
     MongoDB does not allow usage of $ sign in dictionary keys, since it is used
     in MongoQL. To fix the issue we encode the query as dict of key/value/operator
     and decode it back upon retrieval.
+  - fix DAS PLY to support value assignment in filters, e.g.
+    block | grep site=T1
 
 - 0.5.3 - 0.5.4 series